requirements.txt
Faker==15.1.3
python-dateutil==2.8.2
six==1.16.0
